Split run_rust into cargo and cranelift modules
